Separate server, db, and stats parser
